requirements.txt
bscscan-python
python-dotenv
sqlalchemy[asyncio] >= 2.0.0
asyncpg